Moving models from db directory to root directory
