+++ services/message/create.go
@@ -8,7 +8,7 @@ import (
 
 	"github.com/castyapp/libcasty-protocol-go/proto"
 	"github.com/castyapp/libcasty-protocol-go/protocol"
-	"github.com/castyapp/grpc.server/db/models"
+	"github.com/castyapp/grpc.server/models"
 	"github.com/castyapp/grpc.server/helpers"
 	"github.com/castyapp/grpc.server/services/auth"
 	"github.com/golang/protobuf/ptypes"
